% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_to_2010.R
\name{back_to_2010}
\alias{back_to_2010}
\title{back_to_2010}
\usage{
back_to_2010(dt = NULL, geoid_name = NULL)
}
\arguments{
\item{dt}{data.table}

\item{geoid_name}{name of the column with the 11-digit census tract FIPS code}
}
\description{
Crosswalks 2011 to 2019 census tract FIPS codes to 2010 census tract FIPS codes.
After 2010, the census tracts defined for the 2010 census changed in some states.
In almost all instances, those changes did not result in different boundaries,
just different FIPS codes (see Appendix 6 to COI 3.0 technical documentation).
This function crosswalks the new FIPS codes that appeared between 2011 and 2019
back to their 2010 FIPS codes. Only works for 2010 census tracts.

The function accepts a data.table as an input and returns a data.table with geoid_new
added, containing the 2010 FIPS codes. The rows need to be uniquely identified by geoid_name.
The data.table cannot contain the following columns: tract, county, tract_stub, tract_copy.
}
\author{
Clemens Noelke
}
