% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geoid.R
\name{check_geoid}
\alias{check_geoid}
\title{check_geoid}
\usage{
check_geoid(dt = NULL, year = NULL)
}
\arguments{
\item{dt}{Data table to be tested. Must contain a column named geoid containing
11-character census tract FIPS codes.}
}
\description{
Checks if dt contains only census tracts as defined for the 2010 (or 2020) census.

This function loads the 2010 (2020) census tract FIPS codes From DDK.TRACTS and checks
if dt contains any census tract FIPS codes that did not exist in 2010 (2020).
The function returns a data table with the following columns: geoid, data1, data2
data1 indicates if an observation was included in the test data table dt
data2 indicates if an observation was included in the reference data table.
If data2 is NA, the observation was not included in the reference data table
and therefore is not a 2010 (2020) census tract.

Will always load the reference data table DDK.TRACTS from SQL database.
}
\author{
Clemens Noelke
}
