% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_push.R
\name{git_push}
\alias{git_push}
\title{git_push}
\usage{
git_push(
  repo_path = NULL,
  commit_message = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{repo_path}{path to the git repository (the repository on your computer)}

\item{commit_message}{the message to attach to the commit}

\item{username}{your github username}

\item{password}{access token generated from github (see: https://github.com/settings/tokens)}
}
\description{
commits and pushes all changes to git repo
}
\author{
brian devoe
}
