% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_dict.R
\name{SQL_dict}
\alias{SQL_dict}
\title{Load table dictionary only}
\usage{
SQL_dict(database = "DDK", table_id = NULL)
}
\arguments{
\item{database}{Name of database to connect to, character vector of length 1. Default is "DDK".}

\item{table_id}{Name of table from SQL database to load into R environment.
See SQL_table function for list of tables.}
}
\description{
Wrapper function that loads the table dictionary only
}
\author{
Clemens Noelke
}
