% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_droptable.R
\name{SQL_droptable}
\alias{SQL_droptable}
\title{SQL_droptable}
\usage{
SQL_droptable(
  database = NULL,
  table_id = NULL,
  user = Sys.getenv("SQL_dba1"),
  password = Sys.getenv("SQL_dba1_pass")
)
}
\arguments{
\item{database}{name of database to write to in SQL database}

\item{table_id}{name the table that will appear in SQL database}

\item{user}{username for SQL database, this function will only work with administrator credintials}

\item{password}{password for SQL database administrator}
}
\description{
drops table from a give database
}
\author{
brian devoe
}
