% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_kill_connection.R
\name{SQL_kill_connection}
\alias{SQL_kill_connection}
\title{SQL_kill_connection}
\usage{
SQL_kill_connection(user = NULL, password = NULL)
}
\description{
kills processes from a database
}
\author{
brian devoe
}
