% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_load.R
\name{SQL_load}
\alias{SQL_load}
\title{Load table from SQL database}
\usage{
SQL_load(
  database = "DDK",
  table_id = NULL,
  columns = NULL,
  filter = NULL,
  noisily = TRUE,
  show_metadata_only = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{database}{Name of database to connect to, character vector of length 1. Default is "DDK".}

\item{table_id}{Name of table from SQL database to load into R environment.
See SQL_table function for list of tables.}

\item{columns}{Columns to load from table. Default is all columns. Input format is a vector, i.e.
columns = c("col1", "col2", "col3", ...)}

\item{filter}{Filter to apply to table. Default is no filter. Input format is a string (DO NOT USE)
filter = c("col1 = 'value1', col2 = 'value2', col3 >= 'value3', ...)}

\item{noisily}{Print out dictionary and metadata of table. Default is TRUE}

\item{show_metadata_only}{Default is FALSE. If TRUE, function call only prints out dictionary and metadata.}

\item{overwrite}{Default is FALSE. If FALSE and table exists in /data/source_data/sql/ folder, function
will load table from Rdata file. If TRUE, function will download table from SQL database and
overwrite the existing Rdata file.}
}
\description{
Wrapper function for SQL SELECT queries from the SQL database.
By default, it saves table as Rdata file in the HOME/data/source_data/sql/ folder. If the
table already exists, it will load the table from the Rdata file, unless
overwrite is set to TRUE (FALSE by default). If overwrite is set to TRUE, the function
will load the table from the SQL database and overwrite the existing Rdata file. The
function needs the HOME vector to exist in the global environment and point to the
Git root directory.
}
\author{
brian devoe
}
