% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_table_id_list.R
\name{SQL_table_id_list}
\alias{SQL_table_id_list}
\title{SQL_table_id_list}
\usage{
SQL_table_id_list(database = NULL)
}
\arguments{
\item{database}{Name of database to connect to. Default is 'coi'.}
}
\description{
Call function to list available tables in COI SQL database
}
\author{
brian devoe
}
