% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_table_names.R
\name{SQL_table_names}
\alias{SQL_table_names}
\title{SQL_table_names}
\usage{
SQL_table_names(database = NULL)
}
\arguments{
\item{database}{Name of database to connect to. Default is 'coi'.}
}
\description{
Call function to list available tables in COI SQL database
}
\author{
brian devoe
}
