% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_tables.R
\name{SQL_tables}
\alias{SQL_tables}
\title{SQL_tables}
\usage{
SQL_tables(database = NULL)
}
\arguments{
\item{database}{Name of database to connect to. Default is 'coi'.}
}
\description{
Call function to list available tables in COI SQL database
}
\author{
brian devoe
}
