% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_write.R
\name{SQL_write}
\alias{SQL_write}
\title{SQL_write}
\usage{
SQL_write(
  table = NULL,
  table_id = NULL,
  database = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{table}{data table in R environment to write to SQL database}

\item{table_id}{name the table that will appear in SQL database}

\item{database}{name of database to write to in SQL database}

\item{user}{username for SQL database, this function will only work with administrator credintials}

\item{password}{password for SQL database administrator}
}
\description{
write table from R environment into SQL database
WARNING: This function will not work unless you have database administrator credentials
}
\author{
brian devoe
}
