% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dt.R
\name{subset_dt}
\alias{subset_dt}
\title{subset_dt}
\usage{
subset_dt(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\description{
easy way to subset a data.table by column names without data.table syntax
}
\author{
Clemens Noelke
}
