% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_load.R
\name{SQL_load}
\alias{SQL_load}
\title{SQL_load}
\usage{
SQL_load(
  table = NULL,
  database = NULL,
  columns = NULL,
  dictionary = FALSE,
  metadata = FALSE
)
}
\arguments{
\item{table}{Name of table from SQL database to load into R environment.
See SQL_table function for list of tables.}

\item{database}{Name of database to connect to. Default is 'coi'.}

\item{columns}{Columns to load from table. Default is all columns.
Syntax is 'column1, column2, column3'}

\item{dictionary}{Load dictionary table. Default is FALSE.}

\item{metadata}{Load metadata table. Default is FALSE.}
}
\description{
load table from SQL database
}
\author{
brian devoe
}
