% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_load.R
\name{SQL_load}
\alias{SQL_load}
\title{SQL_load}
\usage{
SQL_load(table = NULL, database = NULL, columns = NULL)
}
\arguments{
\item{table}{Name of table from SQL database to load into R environment.
See SQL_table function for list of tables.}

\item{database}{Name of database to connect to. Default is 'coi'.}

\item{columns}{Columns to load from table. Default is all columns. Input format is a vector, i.e.
columns = c("col1", "col2", "col3", ...)}
}
\description{
load table from SQL database
}
\author{
brian devoe
}
