% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_load.R
\name{SQL_load}
\alias{SQL_load}
\title{SQL_load}
\usage{
SQL_load(
  table_id = NULL,
  database = NULL,
  columns = NULL,
  filter = NULL,
  noisily = TRUE
)
}
\arguments{
\item{table_id}{Name of table from SQL database to load into R environment.
See SQL_table function for list of tables.}

\item{database}{Name of database to connect to. Default is 'coi'.}

\item{columns}{Columns to load from table. Default is all columns. Input format is a vector, i.e.
columns = c("col1", "col2", "col3", ...)}

\item{filter}{Filter to apply to table. Default is no filter. Input format is a string (DO NOT USE)
filter = c("col1 = 'value1', col2 = 'value2', col3 >= 'value3', ...)}

\item{noisily}{Print out dictionary and metadata of table. Default is FALSE.}
}
\description{
Wrapper function that helps execute SQL SELECT queries for clients and DDK data team.
}
\examples{
# load full ADI_HIED table from ACS database
ADI_HIED <- SQL_load(table = "ADI_HIED", database = "ACS")
# load selected columns from METRICS_10 table from DDK database
METRICS_10 <- SQL_load(table    = "METRICS_10",
                       database = "DDK",
                       columns  = c("geoid10", "year", "coi30_met"))
# load selected columns and filtered rows from METRICS_10 table from DDK database
METRICS_10 <- SQL_load(table    = "METRICS_10",
                       database = "DDK",
                       columns  = c("geoid10", "year", "coi30_met"),
                       filter   = c("year = 2010"))
}
\author{
brian devoe
}
