% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_tables.R
\name{SQL_tables}
\alias{SQL_tables}
\title{SQL_tables}
\usage{
SQL_tables(database = NULL)
}
\arguments{
\item{database}{Name of database to connect to. Default is 'coi'.}
}
\description{
Returns a table that lists all the given tables in a database. Each row contains the table name
and metadata for that table.
}
\author{
brian devoe
}
