% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_tables.R
\name{SQL_tables}
\alias{SQL_tables}
\title{SQL_tables}
\usage{
SQL_tables(database = "DDK")
}
\arguments{
\item{database}{Name of database to connect to, character vector of length 1. Default is "DDK".}
}
\description{
Returns a data.table that lists all tables in a database. Each row contains the table name and metadata for that table.
}
\author{
brian devoe
}
