% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_write_bulk.R
\name{SQL_write_bulk}
\alias{SQL_write_bulk}
\title{SQL_write_bulk}
\usage{
SQL_write_bulk(
  table = NULL,
  dict = NULL,
  table_id = NULL,
  database = "DDK",
  user = NULL,
  password = NULL,
  overwrite = FALSE,
  test = FALSE
)
}
\arguments{
\item{table}{data.table within R environment to write to SQL database}

\item{dict}{dictionary of table}

\item{table_id}{name of table that will appear in SQL database}

\item{database}{name of database to write to in SQL database -- default is 'DDK'}

\item{overwrite}{logical to overwrite table if it exists -- default is FALSE}

\item{test}{logical to test function by checking if original data.frame is identical to data loaded
into SQL server -- default is FALSE}
}
\description{
write table from given directory to COI SQL database using bulk/batch method
WARNING: will throw and error if local infile is not enabled as root user use 'SET GLOBAL local_infile = 1;' to enable
}
\author{
brian devoe
}
